+++ awx/main/access.py
@@ -1420,7 +1420,7 @@ class JobAccess(BaseAccess):
                 if fd == 'extra_credentials':
                     job_fields[fd] = job_extra_credentials
                 job_fields[fd] = getattr(obj, fd)
-            accepted_fields, ignored_fields = obj.job_template._accept_or_ignore_job_kwargs(**job_fields)
+            accepted_fields, ignored_fields, errors = obj.job_template._accept_or_ignore_job_kwargs(**job_fields)
             # Check if job fields are not allowed by current _on_launch settings
             for fd in ignored_fields:
                 if fd == 'extra_vars':